Update: Code duplication reduced
