+++ restapi/utils.py
@@ -78,3 +78,26 @@ def multi_threaded_reader(urls, num_threads):
             result.extend(data.split("\n"))
     result = sorted(result, key=lambda elem:elem[1])
     return result
+
+def normalize(expense):
+    user_balances = expense.users.all()
+    dues = {}
+    for user_balance in user_balances:
+        dues[user_balance.user] = dues.get(user_balance.user, 0) + user_balance.amount_lent \
+                                  - user_balance.amount_owed
+    dues = [(k, v) for k, v in sorted(dues.items(), key=lambda item: item[1])]
+    start = 0
+    end = len(dues) - 1
+    balances = []
+    while start < end:
+        amount = min(abs(dues[start][1]), abs(dues[end][1]))
+        user_balance = {"from_user": dues[start][0].id, "to_user": dues[end][0].id, "amount": amount}
+        balances.append(user_balance)
+        dues[start] = (dues[start][0], dues[start][1] + amount)
+        dues[end] = (dues[end][0], dues[end][1] - amount)
+        if dues[start][1] == 0:
+            start += 1
+        else:
+            end -= 1
+    
+    return balances
